# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateReportDto": {
        "properties": {},
        "type": "object"
      },
...
      "ErrorBodyDto": {
        "properties": {
          "code": {
            "example": "VALIDATION_ERROR",
            "type": "string"
          },
          "currentReportId": {
            "description": "REPORT_IN_PROGRESS only",
            "format": "uuid",
            "nullable": true,
            "type": "object"
          },
          "currentVersion": {
            "description": "VERSION_CONFLICT only",
            "type": "number"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            },
            "type": "array"
          },
          "message": {
            "example": "Please correct the highlighted fields.",
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
      "ErrorEnvelopeDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "FieldErrorDto": {
        "properties": {
          "code": {
            "example": "DECIMAL_SCALE_EXCEEDED",
            "type": "string"
          },
          "field": {
            "example": "salary",
            "type": "string"
          },
          "message": {
            "example": "Salary must have at most 2 decimal places.",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "MetaDto": {
        "properties": {
          "requestId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
      "NarrativeDto": {
        "properties": {
          "bullets": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "headline": {
            "type": "string"
          }
        },
        "required": [
          "headline",
          "bullets"
        ],
        "type": "object"
      },
      "PageMetaDto": {
        "properties": {
          "page": {
            "example": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 20,
            "type": "number"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "total": {
            "example": 5,
            "type": "number"
          },
          "totalPages": {
            "description": "0 when total is 0",
            "example": 1,
            "type": "number"
          }
        },
        "required": [
          "requestId",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
...
      "ReportDetailDto": {
        "properties": {
          "attempts": {
            "type": "number"
          },
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "errorCode": {
            "nullable": true,
            "type": "string"
          },
          "generatedBy": {
            "enum": [
              "GEMINI",
              "TEMPLATE"
            ],
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "model": {
            "nullable": true,
            "type": "string"
          },
          "narrative": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NarrativeDto"
              }
            ],
            "nullable": true,
            "type": "object"
          },
          "promptVersion": {
            "type": "string"
          },
          "snapshot": {
            "$ref": "#/components/schemas/SnapshotDto"
          },
          "snapshotCapturedAt": {
            "format": "date-time",
            "type": "string"
          },
          "source": {
            "enum": [
              "MANUAL",
              "SCHEDULED"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "type": "string"
          },
          "totalEmployees": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "status",
          "source",
          "snapshotCapturedAt",
          "totalEmployees",
          "createdAt",
          "completedAt",
          "generatedBy",
          "model",
          "promptVersion",
          "errorCode",
          "snapshot",
          "narrative",
          "attempts"
        ],
        "type": "object"
      },
      "ReportSummaryDto": {
        "properties": {
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "errorCode": {
            "nullable": true,
            "type": "string"
          },
          "generatedBy": {
            "enum": [
              "GEMINI",
              "TEMPLATE"
            ],
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "model": {
            "nullable": true,
            "type": "string"
          },
          "promptVersion": {
            "type": "string"
          },
          "snapshotCapturedAt": {
            "format": "date-time",
            "type": "string"
          },
          "source": {
            "enum": [
              "MANUAL",
              "SCHEDULED"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "type": "string"
          },
          "totalEmployees": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "status",
          "source",
          "snapshotCapturedAt",
          "totalEmployees",
          "createdAt",
          "completedAt",
          "generatedBy",
          "model",
          "promptVersion",
          "errorCode"
        ],
        "type": "object"
      },
...
      "SnapshotDepartmentDto": {
        "properties": {
          "active": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "inactive": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "total",
          "active",
          "inactive"
        ],
        "type": "object"
      },
      "SnapshotDto": {
        "properties": {
          "activeEmployees": {
            "type": "number"
          },
          "businessDate": {
            "format": "date",
            "type": "string"
          },
          "capturedAt": {
            "format": "date-time",
            "type": "string"
          },
          "departments": {
            "items": {
              "$ref": "#/components/schemas/SnapshotDepartmentDto"
            },
            "type": "array"
          },
          "inactiveEmployees": {
            "type": "number"
          },
          "schemaVersion": {
            "enum": [
              1
            ],
            "type": "number"
          },
          "timezone": {
            "enum": [
              "Asia/Bangkok"
            ],
            "type": "string"
          },
          "totalEmployees": {
            "type": "number"
          }
        },
        "required": [
          "schemaVersion",
          "capturedAt",
          "businessDate",
          "timezone",
          "totalEmployees",
          "activeEmployees",
          "inactiveEmployees",
          "departments"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "employee_console.local.sid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "NestJS API for Employee Console. Browser clients authenticate with the session cookie (Google OIDC) and send X-CSRF-Token on mutations. /internal/v1 is for the n8n worker/scheduler (bearer tokens).",
    "title": "Employee Console API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reports": {
      "get": {
        "operationId": "Reports_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "QUEUED",
                "RUNNING",
                "SUCCEEDED",
                "FAILED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/ReportSummaryDto"
                      },
                      "type": "array"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Newest first (createdAt desc, id desc)",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN / ACCOUNT_NOT_ALLOWED / CSRF_INVALID",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "RATE_LIMITED / REPORT_QUOTA_EXCEEDED",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      },
      "post": {
        "operationId": "Reports_create",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReportSummaryDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Queued; poll GET /api/v1/reports/{id}",
            "headers": {
              "Idempotency-Replayed": {
                "description": "\"true\" when the response was replayed for a repeated Idempotency-Key",
                "schema": {
                  "type": "string"
                }
              },
              "Location": {
                "description": "Path of the created resource",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VALIDATION_ERROR / INVALID_QUERY / INVALID_IDEMPOTENCY_KEY / MALFORMED_JSON",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN / ACCOUNT_NOT_ALLOWED / CSRF_INVALID",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "VERSION_CONFLICT / IDEMPOTENCY_CONFLICT / REPORT_IN_PROGRESS / STALE_LEASE / REPORT_ALREADY_COMPLETED",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "RATE_LIMITED / REPORT_QUOTA_EXCEEDED",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "AI_NOT_CONFIGURED / AUTH_NOT_CONFIGURED / DEPENDENCY_UNAVAILABLE",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{id}": {
      "get": {
        "operationId": "Reports_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ReportDetailDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/MetaDto"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "UNAUTHENTICATED / SESSION_EXPIRED / INVALID_SERVICE_TOKEN",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "FORBIDDEN / ACCOUNT_NOT_ALLOWED / CSRF_INVALID",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "EMPLOYEE_NOT_FOUND / REPORT_NOT_FOUND",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "RATE_LIMITED / REPORT_QUOTA_EXCEEDED",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            },
            "description": "INTERNAL_ERROR",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated UUID for this request",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "servers": [],
...
}